feat: change custom footer logic
