+++ src/components/footers/footer-links/footer-links.tsx
@@ -1,17 +1,12 @@
 import classes from "./footer-links.module.scss"
 import { FooterLinksProps } from "./footer-links.types"
-import { ActionIcon, Container, Group, Text, rem } from "@mantine/core"
-import {
-  IconBrandInstagram,
-  IconBrandTwitter,
-  IconBrandYoutube,
-} from "@tabler/icons-react"
+import { Container, Group, Text } from "@mantine/core"
 
 export function FooterLinks({
-  footerLabel = "Build fully functional accessible web applications faster than ever",
-  yearLabel = "© 2020 mantine.dev. All rights reserved.",
+  copyRightContent,
+  leftContent,
   linksData,
-  logo,
+  socialContent,
 }: FooterLinksProps) {
   const groups = linksData.map((group) => {
     const links = group.links.map((link, index) => (
@@ -37,18 +32,11 @@ export function FooterLinks({
   return (
     <footer className={classes.footer}>
       <Container className={classes.inner}>
-        <div className={classes.logo}>
-          {logo}
-          <Text size="xs" c="dimmed" className={classes.description}>
-            {footerLabel}
-          </Text>
-        </div>
+        <div className={classes.logo}>{leftContent}</div>
         <div className={classes.groups}>{groups}</div>
       </Container>
       <Container className={classes.afterFooter}>
-        <Text c="dimmed" size="sm">
-          {yearLabel}
-        </Text>
+        {copyRightContent}
 
         <Group
           gap={0}
@@ -56,24 +44,7 @@ export function FooterLinks({
           justify="flex-end"
           wrap="nowrap"
         >
-          <ActionIcon size="lg" color="gray" variant="subtle">
-            <IconBrandTwitter
-              style={{ width: rem(18), height: rem(18) }}
-              stroke={1.5}
-            />
-          </ActionIcon>
-          <ActionIcon size="lg" color="gray" variant="subtle">
-            <IconBrandYoutube
-              style={{ width: rem(18), height: rem(18) }}
-              stroke={1.5}
-            />
-          </ActionIcon>
-          <ActionIcon size="lg" color="gray" variant="subtle">
-            <IconBrandInstagram
-              style={{ width: rem(18), height: rem(18) }}
-              stroke={1.5}
-            />
-          </ActionIcon>
+          {socialContent}
         </Group>
       </Container>
     </footer>
